chore: 상금 상수 위치 변경 및 파일명 변경
