+++ src/constants/LottoSystem.js
@@ -10,3 +10,11 @@ export const LOTTO_SYSTEM = Object.freeze({
   FOUR_MATCH: 4,
   THREE_MATCH: 3,
 });
+
+export const PRIZE_OF_MATCH_COUNT = Object.freeze({
+  [LOTTO_SYSTEM.SIX_MATCH]: 2000000000,
+  [LOTTO_SYSTEM.FIVE_WITH_BONUS_MATCH_IDX]: 30000000,
+  [LOTTO_SYSTEM.FIVE_MATCH]: 1500000,
+  [LOTTO_SYSTEM.FOUR_MATCH]: 50000,
+  [LOTTO_SYSTEM.THREE_MATCH]: 5000,
+});
