+++ src/domain/calculateRevenue.js
@@ -1,4 +1,4 @@
-import { LOTTO_SYSTEM } from '../constants/messages.js';
+import { LOTTO_SYSTEM, PRIZE_OF_MATCH_COUNT } from '../constants/LottoSystem.js';
 
 export const calculateRevenue = (matchCounts, purchasePrice) => {
   const totalRevenue = matchCounts
@@ -8,14 +8,6 @@ export const calculateRevenue = (matchCounts, purchasePrice) => {
 
   return Number(((totalRevenue / purchasePrice) * 100).toFixed(1));
 };
-
-export const PRIZE_OF_MATCH_COUNT = Object.freeze({
-  [LOTTO_SYSTEM.SIX_MATCH]: 2000000000,
-  [LOTTO_SYSTEM.FIVE_WITH_BONUS_MATCH_IDX]: 30000000,
-  [LOTTO_SYSTEM.FIVE_MATCH]: 1500000,
-  [LOTTO_SYSTEM.FOUR_MATCH]: 50000,
-  [LOTTO_SYSTEM.THREE_MATCH]: 5000,
-});
 
 export const calculateRevenueByMatch = (matchCount) => {
   return PRIZE_OF_MATCH_COUNT[matchCount] || 0;
